requirements.txt
Flask==1.1.1
PyMySQL==0.9.3
Flask-Migrate==2.5.2
Flask-SQLAlchemy==2.4.1
python-dotenv==0.11.0
